+++ Evergreen/Public/Get-AdobeBrackets.ps1
@@ -19,7 +19,6 @@ Function Get-AdobeBrackets {
             Description:
             Returns the latest Adobe Brackets version number and download.
     #>
-    [Alias("Get-Atom")]
     [OutputType([System.Management.Automation.PSObject])]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSUseSingularNouns", "")]
     [CmdletBinding()]