+++ Evergreen/Public/Get-MicrosoftOneDrive.ps1
@@ -52,21 +52,21 @@
 
                     # Construct the output for EXE; Return the custom object to the pipeline
                     $PSObject = [PSCustomObject] @{
-                        Version    = $node.currentversion
-                        Ring       = $ring.Name
-                        SHA256hash = $node.binary.sha256hash
-                        Type       = "Exe"
-                        URI        = $node.binary.url
+                        Version = $node.currentversion
+                        Ring    = $ring.Name
+                        Sha256  = $node.binary.sha256hash
+                        Type    = "Exe"
+                        URI     = $node.binary.url
                     }
                     Write-Output -InputObject $PSObject
 
                     # Construct the output for MSIX; Return the custom object to the pipeline
                     $PSObject = [PSCustomObject] @{
-                        Version    = $node.currentversion
-                        Ring       = $ring.Name
-                        SHA256hash = If ($node.msixbinary.sha256hash) { $node.msixbinary.sha256hash } Else { "N/A" }
-                        Type       = "Msix"
-                        URI        = $node.msixbinary.url
+                        Version = $node.currentversion
+                        Ring    = $ring.Name
+                        Sha256  = If ($node.msixbinary.sha256hash) { $node.msixbinary.sha256hash } Else { "N/A" }
+                        Type    = "Msix"
+                        URI     = $node.msixbinary.url
                     }
                     Write-Output -InputObject $PSObject
                 }